Update CI - build per platform, update images, remove GPDB4 and SLES
With the addition of go-sqlite3 and CGO, we now link to libc libraries during our build.  Static
linking was determined to be unviable for our project, so we now must build an artifact for each of
our supported platforms (RHEL6, RHEL7, RHEL8, and Ubuntu).  This required some significant overhauls
to our CI.

We are discontinuing the publication and release of gpbackup artifacts for
the SLES platform.  All tests and builds for it are removed from our CI.

We are discontinuing the publication and release of gpbackup artifacts for
GPDB4.  All tests and builds for it are removed from our CI.

Finally, we also want to update our CI to use modern images (Rocky8) for all generic
containers. These are those used for generic build/tarring, and as jump-boxes to cluster VMS.