Update all dependencies for Go 1.19
The dependencies for this project were fairly out-of-date, not because
it relied on anything specific in older versions of any dependencies but
simply because they hadn't been updated in a while.

As part of a general effort to update GPDB Go projects to 1.19, this
commit updates the required Go version and updates all dependencies to
their currently-latest versions.

Additionally, now that gpbackup is using the same version of Go as the
CI images, installation of Go during test setup is no longer required.