Add migrate-history command to backward-compatibility e2e test
The purpose of this test is to use an older version of gpbackup to create backups, and then a newer
version to restore them.  This workflow will now require an invocation of gpbackup_manager to
migrate the legacy history file to the new sqlitedb format.  So, add one.

NOTE: this may require some additional tweaking to ensure that gpbackup_manager is available in the
relevant CI jobs.