Remove foreign key constraint on restore_plan_timestamp
While restore_plan_timestamp should technically always reference the timestamp of a prior backup, in
many circumstances the backup history database may not be complete. Enforcing this relationship
will cause an otherwise valid backup to fail, and it is better to allow it to succeed.  So remove
the constraint.

Also add a TODO to consider logging a warning when such a backup is taken.