Update test coverage script
The syntax and output of "go test cover" has changed since this script
was originally written, so this commit updates those to maintain the
nice formatted script output we expect.

It also adds the rest of the packages with unit test files to the list
of those for which coverage is generated.