Handle panics from goroutines
Our current cleanup mechanisms rely on a recover() call to pick up
FatalOnError.  However, this mechanism does not work if the panic is
called from a separate goroutine.  As such, our cleanup has some gaps
in it that need to be plugged up.

We identify four cases where a panic can be invoked from a goroutine.
In these cases we set up a channel for bubbling the panics up to the
main process, and invoke gplog.Fatal() at the earliest opportunity.

We structure the defers such that waitgroups are still marked Done()
even in case of a panic, so that we can keep the deadlock protection
they offer.

In the future any we will avoid invoking panic() from within goroutines,
so that this handling will not be necessary.