Handle missing SegmentCount for non-resize restore
If we're performing a normal restore of a backup taken with a version
below 1.26.0, the SegmentCount value is missing from the config file and
so it is set to 0.  This can causes problems down the line, so this
commit sets it to be equal to the current cluster size and adds some
tests to catch similar issues in the future.