Improve BackupDataForAllTables
Previous changes in support distributed snapshots for backing up data
introduced bugs in how we handle errors when piping copy commands to
the server. A prior attempt at fixing this was insufficient, so here we
overhaul how the goroutines doing distributed data backup inform of
errors, and how those errors are handled.

Also, some changes are necessary to our integration test suite to
support testing of this code path. Specifically, the distributed code
path is only invoked if there are multiple connections to the cluster
available. So, we update to run all our tests with two connections
open.