Update layout of the gpbackup builds
It seems that Concourse does not enforce isolation between tasks in the
same job.  This means that the build_binaries tasks were having a bug
where the go_components.tar.gz file pulled in for each platform was
non-deterministic, and we were getting GLIBC version errors because the
gppkgs job was, for example, building the RHEL7 gppkg with the RHEL8
binaries.

This commit refactors our build_binaries step so that there is a
separate job for each platform, and the correct binaries will land where
expected.