Refactor ENXIO timeout for restore_helper agents
The timeout added when doing restore can occasionally cause failures of
otherwise successful restores, in cases where the COPY commands from
server lag too far behind the writer opening its half of the pipe.  A
full RCA of the causes and implications of this is not complete, but for
now a smarter timeout helps resolve errors seen in CI.

The scheme put in place is an exponential backoff approach, starting at
10ms and scaling up to about 5m27s, for a total of 15 retries and about
12 minutes.  Logs of the retry delay begin after the 10s wait, to help
keep users informed of what's happening.