Don't fail if unable to write a restore report
In a scenario where a user is restoring a backup in a read-only backup
directory, gprestore will currently fail at the end because it cannot
write the restore report file to that directory.  This is a legitimate
use case, so we should treat this as a warning rather than an error.