Correctly restore tables with inherited generated columns
In GPDB 7, if a table has a generated column and another table inherits
that table, the child table cannot use the GENERATED keyword to define a
default for that column (whether or not it matches that of the parent's
corresponding column) or GPDB will refuse to create the table.

Thus, when backing up such child tables, we need to omit the GENERATED
statement from the definition of any attributes inherited from another
table.