Modify backup directory hierarchy
To simplify resize restores, which require moving backup files between
segment directories in order to match the destination segment, this
commit removes the segment directory from the directory hierarchy that
gpbackup creates when the --backup-dir flag is used, e.g.

    [backup-dir]/gpseg0/backups/[datestamp]/[timestamp]

becomes

    [backup-dir]/backups/[datestamp]/[timestamp]

Backups taken prior to this change can still be restored despite the
change in format.