Add granular COPY progress reporting for GPDB 7
With the introduction of the gp_stat_progress_copy_summary catalog
table, we can now track the number of tuples being copied in or out
while a COPY is in progress, allowing us to provide more detailed
progress reporting during data backup and restore in GPDB 7 and above.

This takes the form of additional progress bars below the main table
progress bar, one for each connection.  We constantly poll the above
catalog table on an available database connection while the backup or
restore is in progress to calculate the completion percentage for the
table currently being copied, using either estimated row counts (on the
backup side) or exact row counts (on the restore side) for each table's
total.

Progress reporting in GPDB 6 and below is unchanged.