Improve gpbackup_manager test coverage
As we have observed some issues with gpbackup_manager that were not
caught by the single-node tests in the pipeline, this commit modifies
the gpbackup_manager test job to have a multi-node setup.

Additionally, the recent change to gpbackup's directory hierarchy
necessitates testing gpbackup_manager with both the old and new
formats, so an additional task is added to run the test suite against a
backup version from before the format change.