Restore perf for GPDB7 with progress bars
The addition of COPY progress bars creates some overhead. While this is
negligible in the case of large tables, it adds substantial runtime when
backing up many smaller tables.  To alleviate this, refactor so that all
operations for COPY progress tracking only trigger after the COPY has
been running for at least 5 seconds.