Always clean up helper processes
We have been seeing zombie helper processes on occasional test runs, and
sometimes even on live systems. There are no obvious goroutines or other
such "permanent hang" risks, and we already have a cleanup routine for
these in case of failure. Call it even for successes, to ensure we don't
leave a mess.