Explicitly create extension schemas when needed
Recent changes to gp_toolkit on GPDB7+ break restores if the gp_toolkit
schema was not already created.  Further we currently hard-fail if an
extension is meant to be in a schema that was not prior created.

To address the gp_toolkit issue and provide some additional stability,
explicitly add a CREATE IF NOT EXISTS statement for extension schemas.