Remove commented regex for file checking
To avoid cluttering the code, remove the below comment. This was the
regex used for checking for correct file counts before the
--single-backup-dir structure was added to our code base.  We're
dropping it in a very loud and clean way to make it easy to find in case
it is needed for investigation around possible performance regressions.

The comment in question:
// this is the original regex from default-dir structure. It is kept here commented,
// for use in possible future perf investigations if needed
// return fmt.Sprintf("find %s -type f | wc -l", globalFPInfo.GetDirForContent(contentID))