Move constraints back to postdata for data load performance
Originally, constraints were dumped in predata. This caused a restore
issue with tables that used constraints that used the `NOT VALID`
clause. This is because there were scenarios where there was data
grandfathered into tables that violate the constraint that has not yet
been resolved by users, but allowed because of the `NOT VALID` clause.
Because of this issue, and the assumption that nothing can depend on
constraints, constraints were moved from predata to postdata to allow
tables with data that violate `NOT VALID` constraints to restore.
https://github.com/greenplum-db/gpbackup/commit/90b3b54b83fe01f8107a329f2bd10652abd67bd6

It was then later discovered that views can depend on constraints. All
constraints, expect ones with `NOT VALID` clause, were moved back into
predata.
https://github.com/greenplum-db/gpbackup/commit/5895076a2678b00e50455e889d7df93d003fd3e8

While this fixes the dependency issues, this is a significant
performance regression because it is much more performant to restore
constraints after all data has been restored. This is because
constraints use internal indexes. If constraints are restored during
predata, data load would slow down considerably because every row insert
would trigger constraint index update.

In order resolve both the performance regression and original issue, we
can move constraint restore back into postdata and resolve the circular
dependency of a predata view depending on a postdata constraint using a
dummy view in predata. Now if a view that depends on a constraint is
detected, a dummy view that satisfies dependency conditions for other
relations gets dumped in predata. The dummy view and will eventually get
replaced by the dump of the full view definition in postdata after
constraints are dumped.

This circular dependency of predata views depending on postdata
constraints does not exist on GPDB5. View dependency on constraints
exist because of the feature to allow incomplete GROUP BY lists that was
introduced in GPDB6 (postgres 9.1). This commit allows the parser to
recognize functional dependency on primary keys. This allows a table's
other columns to be referenced without listing them in GROUP BY, so long
as the primary key column(s) are listed in GROUP BY.
https://github.com/greenplum-db/gpdb/commit/e49ae8d3bc588294d07ce1a1272b31718cfca5ef

Co-authored-by: Karen Huddleston <khuddleston@vmware.com>