Revert "Fix some bugs in sections code"
This reverts commit 32effa35afa792963acf8dd55f716e84070765ee.

Recent feature releases like the 7x granular progress bars and the
parallel metdata restore have made big changes that introduced a number
of regressions to running gpbackup in the background and parallelism in
general. In order to bring gpbackup into a more stable state and bring
the main branch into a releasable state we are reverting features that
make big code changes across the repo so we can vet the changes more
carefully before merging them in.

We are also reverting the "sections" feature so it can be vetted in
depth.