Fix parallel restore of non-sortable statements
The parallel metadata restore implementation was only assigning
valid Tier values to objects that were processed by TopologicalSort,
i.e. statements handling objects with dependencies. Objects not in
this set were assumed by the ExecuteStatements function to have
Tier{0,0}. There was no distinction between objects that could have
a Tier set but it was missing (backups prior to parallel metadata
restore feature), or objects that never are assigned a tier
(ANALYZE, CREATE INDEX, etc.).

The TopologicalSort comment states A tier[0]==0 value is used to
indicate that this object must be restored serially and should not
be parallelized. Because all metadata commands flow through the same
execution functions, every statement not assigned a valid Tier >
{0,0} would always execute in serial, regardless of the jobs flag.

The fix provided in this commit is short-term, but the performance
impact of this regression is critical enough to require an
immediate fix. The original ExecuteStatements function has been
brought back, which doesn't split statements based on Tier. The
Tier specific case has been moved into a new function
ExecutePredataStatements. This results in some unfortunate code
duplication that will need to be remedied. A 'section' argument
has been added to ExecuteRestoreMetadataStatements, which is
used to determine which of the Execute functions to run.
Currently, passing "predata" will execute ExecutePredataStatements,
while any other option executes ExecuteStatements.

The commit also refactors out the ExecuteStatementsAndCreateProgressBar
so there aren't quite so many layers of function invocations.

Also it fixes some off-by-one issues with connections that are a result
of recent reverts. The restore side does not create an extra
communication connection, while the backup side does. This leads
to a not so subtle footgun when handling the parallel workers. The
connection pool and worker code should be unified for backup and
restore.

Also of note, because scheduleStatementsOnWorkers assigns all
statements up front before passing them into the worker channels,
there is likely to be processing skew on certain workers. Before,
workers simply grabbed the next free statement from the queue.
This limitation further dampens some of the performance gains of parallel
metadata restore.