Partial revert "Fix cleanup calls in plugin_test.sh"
This is a partial revert of commit fc807b0cceb4b3e98599442c49ce53f998732d05.

fc807b0cc is adding an undocumented argument to the plugin_test.sh.
This test is provided to third parties to aid in their plugin
development. The test is written to exercise and test a plugin's
functionality. fc807b0cc introduces changes in an effort to get the
test to clean up after itself, but also doesn't seem to fully work as
mention in a new comment.

This commit also introduces utility/debugging functions we will keep to
maintain parity with ddboost and s3 plugin. The API version will also be
kept at 0.5.0.