Revamp restore pipe logic to use unique pipe names with batch number
This big change is to fix several race condition bugs when using big to
small resize restore that result in the same scenario. A deadlocked
gprestore.

Context:
To complete big to small restore, data for segments that do not exist in
the new cluster must be loaded onto different segment. Method of
choosing segment is round robin. This results in multiple contentIDs of
a backup being restored to a segment during restore. This grouping of
backup content Ids to a segment is a batch.

Previously, the resize restore pipe logic would reuse pipes multiple
times to load batches for a table.  This likely resulted in a few hanging
bugs because the same pipe was being deleted and recreated for each
contentID in a batch. This behavior of deleting and recreating the pipe
that gets reused breaks the following assumption.

Assumption:
By the time the current COPY starts data load, gprestore expects the
next COPY pipe to be ready to read from.

visual example:
gprestore        pipes         helper
COPY 0 runnning  pipe0         created pipe1 before pipe0 is written to
COPY 1           pipe1 exists
COPY 2

Next pipe is pre-created by restore_helper before the current pipe is
used because without this bevavior, the next gprestore COPY may
immediately fail due to missing pipe. When reusing the same pipe for a
batch the "next pipe" is the same as the current pipe being used and
can't be pre-created. This may lead to a number of race conditions.
1. COPY1 reader connects to pipe, restore helper recreates reused pipe,
   COPY1 gets disconnected, errors and moves on to COPY2 (because of
   --on-error-continue), restore_helper now waits for COPY1 to connect
   and now we have a deadlock.
2. COPY1 reader connects to pipe, restore is in the middle of
   recreating the reused pipe and it does not exist yet. COPY1 fails due
   to missing pipe, and moves onto COPY2, restore_helper finished pipe
   recreate and waits for COPY1 to connect and results in a deadlock.

By making the pipes unique to include a batch element number, every COPY
gets a unique pipe. Restore helper can guarantee the next pipe is always
ready before using the current pipe, and we can avoid the race
conditions listed that result in deadlock.

How unique pipe names was implemented:
The restore helper oid list is now populated with an oid and a batch
number. An oid may no be listed multiple times with a different batch
number. Each entry in the oid list represents a COPY restore will run.
Oid and batch number is by restore_helper to create a unique pipe name
for each copy. This change results in simpler restore and helper loop
logic. Replicated tables logic is now also handled during the creation
of the oid list.

While implementing this fix it is determined that we should not skip the
rest of a batch's COPYs if an error is encountered. Skipping COPYs is
difficult because we'd have to have an external mechanism to signal
restore_helper to skip a restore loop. This sounds dangerous and if done
wrong, seem like it would always result in a deadlock. To make gprestore
and helper synchronization easier to maintain, reason about, and less
error prone always run every COPY even if we know it may fail.

Another improvement that was made is that replicated tables are
preprocessed while creating the oid list. We only want to load
replicates once and we can control that by only having oids for
replicated tables written once into the oid list with batch 0. This
simplifies restore_helper because we do not need replicated table file
and special replicated table logic in its restore loop.

The helper process for gpbackup is untouched.

Co-authored-by: Karen Huddleston <khuddleston@vmware.com>
Co-authored-by: Kevin Yeap <kyeap@vmware.com>