Check for agent errors before a copy happens
In the past we've seen gpbackup_helper fail to exit cleaning and die
while leaving pipes around. This would cause gprestore to hand even if
an error file was made because checking for error file happens after a
COPY is issued. But this sequence of events would result in COPY waiting
for a restore helper that is dead.

Co-authored-by: Karen Huddleston <khuddleston@vmware.com>
Co-authored-by: Kevin Yeap <kyeap@vmware.com>
Co-authored-by: Brent Doil <bdoil@vmware.com>