Always write nil byte before flushing pipe and log error file creation
To help prevent hanging COPYs, flush will always a nil byte before
closing a pipe to disconnect COPY readers from a pipe in case there are
any connected. This is likely not needed because experiments show that
readers are disconnected when a pipe is deleted even when nothing is
written to the pipe, but it is better to be safe here.

If an restore_helper encounters an error while writing the error file,
it is worth logging the error. Since this error file is how
restore_helper signals to gprestore that it encountered an irrecoverable
error, we want to know if from restore_helper itself if it failed to
signal to gprestore that it has died.

Co-authored-by: Karen Huddleston <khuddleston@vmware.com>
Co-authored-by: Kevin Yeap <kyeap@vmware.com>
Co-authored-by: Brent Doil <bdoil@vmware.com>