Move some logs to Info or Progress so they will always show
Now that we are making our default log file verbosity INFO we still want to
make sure we have enough information to debug issues on the default log level
if they come up. Introduce new command gplog.Progress to be used with progress
bars. It will log table information to the file but not to the terminal so
progress bar output still looks good. It also logs at Info level. Also allow
for multiple log levels in the helpers so we will only see the most important
output by default.