Only do one batch for replicated tables on gprestore side
We already control the batches for replicated tables on the helper side, but we
were not doing the same on the gprestore side. We ended up calling COPY for
batches that the helper was not doing.

This was not showing up as an error in CI until we removed the extra `cat -`.
We think this was because it covered the `No such file` error and `cat -` would
exit immediately with a 0 return code so COPY did not report it as an error. It
was not loading any extra data, just running unnecessary commands.