Check wasTerminated more and wait for cleanup to finish always
There could be a race condition where backup or restore is terminated during
Teardown so the process completes and exits without waiting for Cleanup to
finish. Also check if wasTerminated and return a few more places so we won't
start new COPY processes for example

Also add FlakeAttempts back to one tests that sometimes flakes. The backup one
already ha sit so should be ok to add to restore too.